Add words in upgradeneeded event to fix error and refactor
