+++ src/data/api.ts
@@ -1,19 +1,23 @@
-import type { WordEntries } from "@/types";
-import { localDB } from "./indexedbd.ts/indexeddb";
+import { localDB } from "./indexedbd";
 import { cloudflareApi } from "./cloudflare-api";
 
 const API_URL = './api/dictionary';
 
-export async function getKeys() {
+export const api = {
+    getKeys,
+    getWordEntries,
+}
+
+async function getKeys() {
     try {
         const keys: string[] = await fetch(`${API_URL}/keys`).then(res => res.json());
         return keys;
     } catch (error) {
-        console.log(error);
+        console.error(error);
         return [];
     }
 }
-export async function getWordEntries(word: string) {
+async function getWordEntries(word: string) {
     try {
         const entries = await localDB.getWordEntries(word);
         if (!entries) {
@@ -23,16 +27,5 @@ export async function getWordEntries(word: string) {
     } catch (error) {
         const entries = await cloudflareApi.getWordEntries(word)
         return entries;
-    }
-}
-
-
-export async function geAllWordEntries() {
-    try {
-        const entries: WordEntries = await fetch(`${API_URL}/all`).then(res => res.json());
-        return entries;
-    } catch (error) {
-        console.log(error);
-        return [];
     }
 }
