+++ src/main.ts
@@ -4,7 +4,6 @@ import router from './router'
 import '@fontsource-variable/nunito/wght.css';
 import './assets/styles/styles.scss';
 import { createHead } from '@unhead/vue';
-import { localDB } from './data/indexedbd.ts/indexeddb';
 
 const app = createApp(App)
 app.use(router)
@@ -13,5 +12,3 @@ const head = createHead()
 app.use(head)
 
 app.mount('#app')
-
-localDB.init();
